==> R/2_glucose.Rmd <==
---
title: "Glucose"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Import

```{r}
library(data.table)
library(ggplot2)
library(scales)

## data ## ---------------------------------------------------------------------

glucose <- fread("../data/analysis/glucose.csv")
glucose[time >= 0,n_icu := seq(1,.N),by=patientunitstayid]
glucose[time >= 0,n_icu := seq(1,.N),by=patientunitstayid]
glucose[,dy25 := floor(dy/0.25)*0.25]
glucose_bedside <- glucose[!is.na(glucose_bedside)]

icustay <- fread("../data/raw/icustay.csv")
```

# Table 1

```{r}
##### table 1 #####
# n encounters
length(unique(glucose[time >=0,patientunitstayid]))
length(unique(glucose[time >=0 & hosp_mort == 0,patientunitstayid]))
length(unique(glucose[time >=0 & hosp_mort == 1,patientunitstayid]))

# bedside measures per day
tab <- glucose_bedside[time >= 0 & icu_los_hours > 0 & time < icu_los_hours*60,
                       .(.N,time_diff = mean(time_diff,na.rm=TRUE),
                         diabetes = max(diabetes),
                         day_1_hyper = max(day_1_hyper),
                         day_1_hypo = max(day_1_hypo),
                         hosp_mort = max(hosp_mort),
                         insulin = max(insulin),
                         insulin_infusion = max(insulin_infusion),
                         discharge_time = mean(unitdischargeoffset/60)),
                       by=.(patientunitstayid,icu_los_hours)]
tab[,N_per_24 := N/icu_los_hours*24]
summary(tab[,N_per_24])
summary(tab[hosp_mort == 0,N_per_24])
summary(tab[hosp_mort == 1,N_per_24])

# admission blood glucose
glucose[n_icu == 1,summary(glucose)]
glucose[n_icu == 1 & hosp_mort == 0,summary(glucose)]
glucose[n_icu == 1 & hosp_mort == 1,summary(glucose)]

# mean bg
tab <- glucose[time >= 0 & time < icu_los_hours*60,.(glucose=mean(glucose),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[,summary(glucose)]
tab[hosp_mort == 0,summary(glucose)]
tab[hosp_mort == 1,summary(glucose)]
t.test(glucose ~ hosp_mort,data=tab)

# mean bg (day 1) > 180
tab <- glucose[time >= 0 & time < 24*60,.(glucose=mean(glucose),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[,.N,by=glucose >= 180]
tab[hosp_mort == 0,.N,by=glucose >= 180]
tab[hosp_mort == 1,.N,by=glucose >= 180]
t.test(glucose ~ hosp_mort,data=tab)

# mean bg (day 1) < 70
tab <- glucose[time >= 0 & time < 24*60,.(glucose=mean(glucose),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[,.N,by=glucose >= 180]
tab[hosp_mort == 0,.N,by=glucose >= 180]
tab[hosp_mort == 1,.N,by=glucose >= 180]
t.test(glucose ~ hosp_mort,data=tab)

# max bg
tab <- glucose[time >= 0 & time < icu_los_hours*60,.(glucose=max(glucose),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[,summary(glucose)]
tab[hosp_mort == 0,summary(glucose)]
tab[hosp_mort == 1,summary(glucose)]

# blood glucose SD
tab <- glucose[time >= 0 & time < icu_los_hours*60,.(glucose=sd(glucose),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[,summary(glucose)]
tab[hosp_mort == 0,summary(glucose)]
tab[hosp_mort == 1,summary(glucose)]
t.test(glucose ~ hosp_mort,data=tab)

# day 1 hypo
tab <- glucose[,.(day_1_hypo=max(day_1_hypo),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[day_1_hypo == 1,.N]
tab[day_1_hypo == 1,.N,by=.(hosp_mort)]

# day 1 hyper
tab <- glucose[,.(day_1_hyper=max(day_1_hyper),hosp_mort=max(hosp_mort)),by=patientunitstayid]
tab[day_1_hyper == 1,.N]
tab[day_1_hyper == 1,.N,by=.(hosp_mort)]
```


# in-text
```{r}
# number measures
glucose[,.N]
glucose[!is.na(glucose_bedside),.N]
glucose[!is.na(glucose_lab),.N]

# number encounters
length(unique(glucose$patientunitstayid))
length(unique(glucose[measure=="bedside"]$patientunitstayid))
length(unique(glucose[measure=="lab"]$patientunitstayid))

# percent stay with a measure
n1 <- icustay[icu_los_hours > 12,.N]
n2 <- icustay[patientunitstayid %in% glucose[time >= 0]$patientunitstayid & icu_los_hours > 12,.N]
n2/n1

# bedside measurements per patient
tab <- glucose_bedside[time >= 0 & icu_los_hours > 0,
                       .(.N,time_diff = mean(time_diff,na.rm=TRUE),
                         diabetes = max(diabetes),
                         day_1_hyper = max(day_1_hyper),
                         day_1_hypo = max(day_1_hypo),
                         insulin = max(insulin),
                         insulin_infusion = max(insulin_infusion),
                         discharge_time = mean(unitdischargeoffset/60)),
                       by=.(patientunitstayid,icu_los_hours)]
tab[,N_per_24 := N/icu_los_hours*24]
summary(tab$N_per_24)
summary(tab[diabetes == 1]$N_per_24)
summary(tab[day_1_hyper == 1]$N_per_24)
summary(tab[day_1_hypo == 1]$N_per_24)
summary(tab[insulin == 1]$N_per_24)
summary(tab[insulin_infusion == 1]$N_per_24)

## glucose quality check -------------------------------------------------------

# number stays in hospitals with no glucose
gluc_hosp <- unique(glucose$hospitalid)
icustay[!hospitalid %in% gluc_hosp,.N]
```

# Graph

```{r}
# samples
set.seed(111)
n_per_patient <- glucose[time > 0,.(.N,icu_los_hours=mean(icu_los_hours)),
                         by=patientunitstayid]
samp10 <- sample(n_per_patient[N > 8 & icu_los_hours > 24 & icu_los_hours < 24*2,
                               patientunitstayid],10)
glucose_samp <- glucose[patientunitstayid %in% samp10]
glucose_samp[,id := paste0("ICU stay - ",LETTERS[as.numeric(factor(patientunitstayid))])]
p1 <- ggplot(glucose_samp,aes(x=time/60,y=glucose)) +
  #geom_smooth(se=FALSE,col="gray50",size=0.5,span = 0.7) +
  geom_line(col="gray") +
  geom_point(aes(col=measure),size=1.0) +
  scale_y_log10(n.breaks=5) +
  scale_color_manual(values=c("black","deepskyblue")) +
  scale_x_continuous(limits=c(0,48.0),breaks = seq(0,48,by=12)) +
  facet_wrap(~id,ncol=2) +
  annotate(geom = "rect",
           fill="pink",
           xmin = 0,
           xmax = 48,
           ymin = 80,
           ymax = 180,
           alpha = 0.25) +
  theme_bw(base_size=14) +
  theme(legend.position = "none") +
  labs(y = "Blood glucose (mg/dL)",x = "Time since ICU admission (hours)")
p1
ggsave("../graphs/glucose_sample.png",p1)
# p23 <- ggplot(glucose,
#        aes(x=glucose,fill=measure)) +
#   scale_x_log10(limits=c(30,1000)) +
#   geom_histogram(col="white") +
#   labs(x = "Blood glucose (mmol/L)",y="Count") +
#   theme_bw(base_size=14)
# p23
# ggsave("graphs/glucose_bed_lab_dist.png",p23)
p2 <- ggplot(glucose,
       aes(x=glucose)) +
  scale_x_log10(limits=c(30,1000)) +
  scale_y_continuous(labels = label_comma()) +
  geom_histogram(col="white",fill="lightblue") +
  labs(x = "Blood glucose (mmol/L)",y="Count") +
  theme_bw(base_size=14)
p2
ggsave("../graphs/glucose_dist.png",p2)
glucose_av <- glucose[,.(glucose = mean(glucose)),by=patientunitstayid]
p3 <- ggplot(glucose_av,
       aes(x=glucose)) +
  scale_x_log10(limits=c(30,1000)) +
  scale_y_continuous(labels = label_comma()) +
  geom_histogram(col="white",fill="lightblue") +
  labs(x = "Mean blood glucose (mmol/L)",y="Count") +
  theme_bw(base_size=14)
p3
ggsave("../graphs/glucose_average_dist.png",p3)
# time between bedside
summary(glucose_bedside$time_diff)
p4 <- ggplot(glucose_bedside[time_diff < 24],aes(x=time_diff)) +
  geom_histogram(binwidth=0.5,fill="lightblue",col="white") +
  scale_y_continuous(labels = label_comma()) +
  scale_x_continuous(breaks = seq(0,12,by=2)) +
  theme_bw(base_size=14) +
  labs(x="Time between measurements (hours)",y="Count") +
  coord_cartesian(xlim=c(0,12))
p4
ggsave("../graphs/glucose_time_measure.png",p4)
# correlation between measures
glucose[,time_diff_all := (shift(time,-1) - time)/60,by=patientunitstayid]
glucose[,time_diff_05 := round(time_diff_all/0.5)*0.5,]
glucose[,glucose_next := shift(glucose,-1),by=patientunitstayid]
corr_res <- list()
for (lag in 1:10) {
  print(lag)
  glucose[,glucose_lag := shift(glucose,-lag),by=patientunitstayid]
  glucose[,time_diff_all := (shift(time,-lag) - time)/60,by=patientunitstayid]
  glucose[,time_diff_05 := round(time_diff_all/0.5)*0.5,]
  glucose_corr <- glucose[time_diff_all < 16,.(.N,corr = cor(glucose,glucose_lag,use = "pairwise.complete.obs")),by=time_diff_05]
  glucose_corr <- glucose_corr[order(time_diff_05)]
  glucose_corr[,lag := lag]
  corr_res[[lag]] <- glucose_corr
}
corr_res <- do.call(rbind,corr_res)
glucose_corr <- corr_res[,.(corr = weighted.mean(corr,N)),by=time_diff_05]
p5 <- ggplot(glucose_corr,aes(x=time_diff_05,y=corr)) +
  geom_smooth(se=FALSE) +
  geom_point() +
  scale_x_continuous(n.breaks = 6) +
  scale_y_continuous(n.breaks = 6) +
  labs(x="Time between measurements (hours)",y="Correlation (Pearson)") +
  theme_bw(base_size=14) +
  coord_cartesian(ylim=c(0,1))
p5
ggsave("../graphs/glucose_corr_time.png",p5)
# population dist over time
q10 <- seq(0.05, 0.95, by = 0.05)
qtab <- glucose[ ,lapply(.SD, quantile, prob = q10, na.rm = TRUE),.SDcols = c("glucose"),by=dy25]
qtab[,quantile := rep(q10,nrow(qtab)/length(q10))]
#qtab <- melt(qtab,id.vars = c("dy25","quantile"))
qtab[,low_high := rep(c(rep("lower",9),"median",rep("higher",9)),nrow(qtab)/length(q10))]
qtab[,group := rep(c(letters[1:9],"median",rev(letters[1:9])),nrow(qtab)/length(q10))]
qtab_lh <- tidyr::pivot_wider(qtab[low_high!="median",.(dy25,low_high,group,glucose)],names_from = "low_high",values_from="glucose")
setDT(qtab_lh)
qtab_m <- qtab[low_high=="median"]
p6 <- ggplot() +
  geom_ribbon(data=qtab_lh,aes(x=dy25,ymin=lower,ymax=higher,group=group),
              alpha=0.2) +
  geom_line(data=qtab_m,aes(x=dy25,y=glucose)) +
  scale_x_continuous(limits=c(0,5)) +
  scale_y_log10(limits=c(70,500),n.breaks=7) +
  labs(x = "Time since ICU admission (days)",y="Blood glucose (mmol/L)") +
  theme_bw(base_size=14)
p6
ggsave("../graphs/glucose_pop_time.png",p6)

# admission glucose and hospital outcome
glucose[time > 0,min_time := min(time),by=patientunitstayid]
glucose[time == min_time]
cuts <- quantile(glucose$glucose,probs=seq(0,1,0.1)[2:9],na.rm=TRUE)
cuts <- unique(round(cuts,-1))
glucose[,glucose_q := cut(glucose,
                          c(0,70,100,140,160,180,200,220,300,Inf),ordered=TRUE)]
hosp_mort_gluc <- glucose[time == min_time & !is.na(glucose_q) & !is.na(diabetes),
                                  .(hosp_mort=mean(hosp_mort,na.rm=TRUE),.N),by=.(glucose_q,diabetes)]

levels(hosp_mort_gluc$glucose_q) <- stringr::str_replace(levels(hosp_mort_gluc$glucose_q),",","-")
levels(hosp_mort_gluc$glucose_q) <- stringr::str_replace(levels(hosp_mort_gluc$glucose_q),"\\(","")
levels(hosp_mort_gluc$glucose_q) <- stringr::str_replace(levels(hosp_mort_gluc$glucose_q),"\\]","")
levels(hosp_mort_gluc$glucose_q) <- stringr::str_replace(levels(hosp_mort_gluc$glucose_q),"-Inf","+")
hosp_mort_gluc[,gluc := as.numeric(glucose_q)]
hosp_mort_gluc[,se := sqrt(hosp_mort*(1-hosp_mort)/N)]
p7 <- ggplot(hosp_mort_gluc,aes(x=glucose_q,y=100*hosp_mort,col=factor(diabetes))) +
  geom_pointrange(aes(ymin=100*(hosp_mort-2*se),ymax=100*(hosp_mort+2*se))) +
  geom_line(aes(x=gluc,alpha=0.2)) +
  labs(x = "Admission blood glucose (mg/dL)",y = "Hospital mortality (%)") +
  scale_color_discrete(name="",labels=c("Non-diabetic","Diabetic")) +
  scale_y_continuous(n.breaks=8,limits=c(0,30)) +
  guides(alpha="none")+
  theme_bw(base_size=14)+
  theme(axis.text.x = element_text(angle=-45,hjust = 0))
p7
ggsave("../graphs/glucose_hos_mort.png",p7)
# data for following tables
tab <- glucose[,.(mean=mean(glucose,na.rm=TRUE),
                  sd=sd(glucose,na.rm=TRUE),
                  hosp_mort = max(hosp_mort),
                  diabetes=max(diabetes),
                  .N),
        by=patientunitstayid]
tab[is.na(diabetes),diabetes := 0]
p8 <- ggplot(tab[N > 3],aes(x=mean,y=sd)) +
  geom_point(alpha=0.05) +
  coord_cartesian(ylim=c(1,500),xlim=c(50,750)) +
  scale_x_log10(n.breaks=7) +
  scale_y_log10(n.breaks=7) +
    labs(x="Mean blood glucose (mg/dL)",y="SD blood glucose (mg/dL)") +
      scale_color_discrete(name="",labels=c("Non-diabetic","Diabetic")) +
  geom_smooth(se=FALSE,aes(col=factor(diabetes)))+
  theme_bw(base_size=14)+
  theme(axis.text.x = element_text(angle=-45,hjust = 0))
p8
tab[,sd_grp := cut(sd,ordered=TRUE,breaks=c(0,5,10,15,20,25,30,40,50,70,100))]
tab1 <- tab[!is.na(sd_grp),.(hosp_mort = mean(hosp_mort,na.rm=TRUE),.N),by=.(sd_grp,diabetes)]
tab1[,se := sqrt(hosp_mort*(1-hosp_mort)/N)]
levels(tab1$sd_grp) <- stringr::str_replace(levels(tab1$sd_grp),",","-")
levels(tab1$sd_grp) <- stringr::str_replace(levels(tab1$sd_grp),"\\(","")
levels(tab1$sd_grp) <- stringr::str_replace(levels(tab1$sd_grp),"\\]","")
levels(tab1$sd_grp) <- stringr::str_replace(levels(tab1$sd_grp),"-Inf","+")
p9 <- ggplot(tab1,aes(x=sd_grp,y=100*hosp_mort,col=factor(diabetes))) +
  geom_pointrange(aes(ymin=100*(hosp_mort-2*se),ymax=100*(hosp_mort+2*se))) +
    scale_color_discrete(name="",labels=c("Non-diabetic","Diabetic")) +
  theme_bw(base_size=14) +
  labs(x = "SD blood glucose (mg/dL)",y="Hospital mortality (%)")+
  theme(axis.text.x = element_text(angle=-45,hjust = 0))
p9
p10 <- ggplot(tab[N > 3 & !is.na(hosp_mort)],aes(x=mean,y=sd,col=factor(hosp_mort))) +
  geom_density_2d() +
  scale_color_discrete(name="Hospital\noutcome",labels=c("Survived","Died")) +
  labs(x="Mean blood glucose (mg/dL)",y="SD blood glucose (mg/dL)") +
  scale_y_log10(n.breaks=7) +
  scale_x_log10(n.breaks=7) +
  theme_bw(base_size=14)+
  theme(axis.text.x = element_text(angle=-45,hjust = 0))
p10

layout_matrix <- matrix(c(1,1,1,6,7,9,
                          2,3,4,5,8,10),ncol=2)
gridExtra::grid.arrange(p1+labs(title="(A)"),
                        p2+labs(title="(B)"),
                        p3+labs(title="(C)"),
                        p4+labs(title="(D)"),
                        p5+labs(title="(F)"),
                        p6+labs(title="(E)"),
                        p7+labs(title="(G)"),
                        p8+labs(title="(H)"),
                        p9+labs(title="(I)"),
                        p10+labs(title="(J)"),
                        layout_matrix=layout_matrix)
p_all <- gridExtra::arrangeGrob(p1+labs(title="(A)"),
                                p2+labs(title="(B)"),
                                p3+labs(title="(C)"),
                                p4+labs(title="(D)"),
                                p5+labs(title="(F)"),
                                p6+labs(title="(E)"),
                                p7+labs(title="(G)"),
                                p8+labs(title="(H)"),
                                p9+labs(title="(I)"),
                                p10+labs(title="(J)"),
                                layout_matrix=layout_matrix)
ggsave("../graphs/glucose.png",p_all,width = 10,height = 16)

```
