==> R/2_insulin.Rmd <==
---
title: "eICU-CRD: Insulin"
output:
  html_document:
    df_print: paged
---

```{r message=FALSE, warning=FALSE}
## packages ##
library(data.table)
library(ggplot2)
library(scales)
library(stringr)
```

# Import data

```{r}
## data ##
insulin_admin <- fread("../data/analysis/insulin_administrations.csv")
insulin_orders <- fread("../data/analysis/insulin_orders.csv")
#insulin_sc <- fread("../data/analysis/insulin_sc.csv")
#insulin_other <- fread("../data/analysis/insulin_other.csv")
icustay <- fread("../data/raw/icustay.csv")
glucose <- fread("../data/analysis/glucose.csv")
glucose[time >= 0,n_icu := seq(1,.N),by=patientunitstayid]
glucose_bedside <- glucose[!is.na(glucose_bedside)]
diabetes <- fread("../data/raw/diabetes.csv")

# subsets
insulin_infusions <- insulin_admin[route=="iv"]
```
```{r}
# new variables
insulin_infusions[,insulin_rate := as.numeric(insulin_rate)]
insulin_infusions[,time_next := shift(time,-1),by=patientunitstayid]
insulin_infusions[,time_gap := time_next - time]
insulin_infusions[,dy0 := floor(time/60/24)]
insulin_infusions[,diabetes_f := fifelse(diabetes == 1,"Diabetic","Non-diabetic")]
```

```{r}
# glucose0
insulin_dy <- unique(c(insulin_orders[time > 0 & time < 24*60,patientunitstayid],
                       insulin_infusions[time > 0 & time < 24*60,patientunitstayid]))
glucose[,insulin_dy1 := ifelse(patientunitstayid %in% insulin_dy,1,0)]
#glucose[,insulin := fifelse(patientunitstayid %in% insulin_pts,1,0)]
glucose[,diabetes_f := fifelse(diabetes == 1,"Diabetic","Non-diabetic")]
```


# Data cleaning

Distribution of insulin rate - set to NA extreme outliers.

```{r}
ggplot(insulin_infusions,aes(x=insulin_rate)) +
  stat_ecdf(geom="step") +
  coord_cartesian(xlim=c(0,100))
```

```{r}
quantile(insulin_infusions$insulin_rate,na.rm = TRUE,probs = seq(0.99,1,0.001))
```

Set rates above 112 to NA

```{r}
insulin_infusions[insulin_rate > 112,insulin_rate := NA]
```

Filter to in-icu for rest of analysis

```{r}
insulin_admin <- insulin_admin[time > 0 & time < icu_los_hours*60]
insulin_orders <- insulin_orders[stop_time > 0 & time < icu_los_hours*60]
insulin_infusions <- insulin_infusions[time_next > 0 & time < icu_los_hours*60]
```

# Hospitals with insulin information

```{r}
insulin_hosps <- unique(c(insulin_orders$hospitalid,insulin_infusions$hospitalid))
print("Num hospitals:")
length(unique(icustay$hospitalid))
print("Num hospitals with any insulin:")
length(insulin_hosps)
print("Num ICU stays in hospitals with no insulin:")
icustay[!hospitalid %in% insulin_hosps,.N]
```

# IV

## Descriptive stats

### ICU stays

ICU stays with IV insulin order:

```{r}
iv_pts <- unique(c(insulin_orders[route=="iv",patientunitstayid],
         insulin_infusions$patientunitstayid))
length(iv_pts)
```

ICU stays with IV (expl. infusion) insulin order:

```{r}
infuse_pts <- unique(c(insulin_infusions$patientunitstayid))
print("IV (non-order):")
length(infuse_pts)
infuse_pts <- unique(c(insulin_orders[route=="iv" & frequency_english == "continuous",patientunitstayid]))
print("IV (order):")
length(infuse_pts)
infuse_pts <- unique(c(insulin_orders[route=="iv" & frequency_english == "continuous",patientunitstayid],
                   insulin_infusions$patientunitstayid))
length(infuse_pts)
```

IV insulin orders by ICU stay outcome

```{r}
print("Survived:")
length(unique(c(insulin_orders[route=="iv" & hosp_mort == 0,patientunitstayid],
                   insulin_infusions[hosp_mort == 0]$patientunitstayid)))
print("Died:")
length(unique(c(insulin_orders[route=="iv" & hosp_mort == 1,patientunitstayid],
                insulin_infusions[hosp_mort == 1]$patientunitstayid)))
```

IV insulin orders route:

```{r}
insulin_orders[,.N,by=route]
```

Num IV insulin orders missing dose:

```{r}
insulin_orders[route == "iv",.N,by=is.na(dose)]
```

### IV info in each table

Source of IV insulin information

```{r}
print("intakeoutput")
insulin_infusions[source_intakeoutput == 1,.N,by=patientunitstayid][,.N]
print("infusion_drug and intakeoutput")
insulin_infusions[source_intakeoutput == 1 & source_infusion_drug == 1,.N,by=patientunitstayid][,.N]
print("infusion_drug only")
insulin_infusions[source_intakeoutput == 0 & source_infusion_drug == 1,.N,by=patientunitstayid][,.N]
print("infusion_drug only")
insulin_infusions[source_infusion_drug == 1,.N,by=patientunitstayid][,.N]
```

### Infusion rate

Heterogeneity in recording infusion rate

```{r}
insulin_infusions[source_infusion_drug == 1,.N,by=tolower(drugname)][order(-N)]
```

```{r}
insulin_infusions[grepl("kg",drugname),insulin_rate := NA]
insulin_infusions[grepl("gtt",drugname),insulin_rate := NA]
```

Insulin infusion rate (all rates)

```{r}
insulin_infusions[source_infusion_drug == 1,summary(insulin_rate)]
```

Insulin infusion rate (units/hr)

```{r}
insulin_infusions[source_infusion_drug == 1 & grepl("units/hr",drugname),summary(insulin_rate)]
```

Number of orders recorded as units/hr

```{r}
insulin_infusions[source_infusion_drug == 1 & grepl("units/hr",drugname),.N]
```

Insulin infusion rate (units/hr) for diabetics

```{r}
insulin_infusions[source_infusion_drug == 1 & diabetes == 1 & grepl("units/hr",drugname),summary(insulin_rate)]
```

### Infusion time

Total time on insulin infusion

```{r}
tab <- insulin_infusions[,.(total=sum(time_gap)/60),by=patientunitstayid]
tab[,summary(total)]
```

### Infusion total

Insulin units per day for those on infusion

```{r}
tab <- insulin_infusions[source_infusion_drug == 1 & drugname == "insulin (units/hr)",
.(patientunitstayid,start_dy=time/60/24,stop_dy=time_next/60/24,time_h = time_gap/60,insulin_rate,diabetes)]
tab[,same_dy := floor(start_dy) == floor(stop_dy)]
tab1 <- tab[same_dy == TRUE,.(total=sum(insulin_rate*time_h)),by=.(patientunitstayid,dy=floor(start_dy),diabetes)]
print("All")
tab1[,summary(total)]
print("Non-diabetics")
tab1[diabetes == 0,summary(total)]
print("Diabetics")
tab1[diabetes == 1,summary(total)]
```

# SC

## Plots

```{r}
hosp_tab <- icustay[,.(Nt=.N),by=.(hospitalid)]
insulin_orders[,usable :=  !is.na(dose) & !(frequency_english %in% c("","na")) & insulin_type != ""]
med_by_hosp <- insulin_orders[,.(usable  = mean(usable),.N),by=.(hospitalid,route)]
med_by_hosp <- merge(med_by_hosp,hosp_tab,by="hospitalid",all.x=TRUE)
labels <- c(iv = "Intravenous", sc = "Subcutaneous")
p1 <- ggplot(med_by_hosp[route != "other"],aes(x=Nt,y=usable)) +
  facet_wrap(~route,labeller=labeller(route = labels)) +
  geom_point() +
  scale_x_continuous(labels = label_comma()) +
  scale_y_continuous(breaks=seq(0,1,0.2),labels=100*seq(0,1,0.2)) +
  labs(x = "Number of ICU stays",y="Complete insulin data (%)") +
  theme_bw(base_size=14)
p1
ggsave(plot=p1,filename="../graphs/complete_insulin_orders_n.png")
# insulin_orders[,rec_dose := fifelse(!is.na(dose),"Dose recorded","No dose recorded")]
# insulin_orders[,rec_freq := fifelse(!is.na(times_per_day),"Frequency recorded","No frequency recorded")]
# insulin_orders[,prn_no := fifelse(prn == "no","Not PRN","PRN")]
# p2 <- ggplot(insulin_orders[route %in% c("sc")],aes(x = prn_no,fill=rec_freq,col=factor(usable))) +
#   geom_bar(position = "dodge",size=1) +
#   facet_wrap(~rec_dose) +
#   scale_fill_discrete(name="") +
#   theme_bw(base_size=14) +
#   scale_y_continuous(labels = label_comma()) +
#   scale_color_manual(values = c(alpha("white",0.0),"black")) +
#   guides(color="none") +
#   labs(x = "",y = "Count") +
#   theme(legend.position = "bottom",
#         legend.box.spacing = unit(0, "pt"),
#         legend.margin=margin(2,2,2,2))
# p2
icustay[,insulin := fifelse(patientunitstayid %in% unique(c(insulin_infusions$patientunitstayid,
                                                            insulin_orders$patientunitstayid)),1,0)]
hospital_insulin <- icustay[,.(insulin=mean(insulin),.N),by=hospitalid]
p3 <- ggplot(hospital_insulin,aes(x=N,y=insulin)) +
  geom_point() +
  scale_y_continuous(breaks=seq(0,1,0.2),labels=100*seq(0,1,0.2)) +
  theme_bw(base_size=14) +
  scale_x_continuous(labels = label_comma()) +
  labs(x = "Number of ICU stays",y = "Patients receiving insulin (%)") +
  geom_hline(yintercept = 0.4,linetype=2)
p3
ggsave(plot=p3,filename="../graphs/insulin_by_hosp.png")
```

## Descriptive stats

### ICU stays

ICU stays with SC insulin order

```{r}
sc_pts <- unique(c(insulin_orders[route=="sc",patientunitstayid],
                 insulin_admin[route=="sc",patientunitstayid]))
length(sc_pts)
```

ICU stays with SC insulin order by outcome

```{r}
# by outcome
print("Survived:")
length(unique(c(insulin_orders[route=="sc" & hosp_mort == 0,patientunitstayid],
              insulin_admin[route=="sc" & hosp_mort == 0,patientunitstayid])))
print("Died:")
length(unique(c(insulin_orders[route=="sc" & hosp_mort == 1,patientunitstayid],
              insulin_admin[route=="sc" & hosp_mort == 1,patientunitstayid])))
```
Is dose/frequency/type information recorded for the SC orders?

```{r}
print("total SC orders:")
insulin_orders[route=="sc",.N]
print("SC orders with dose and frequency and insulin type information:")
insulin_orders[route=="sc" & !is.na(dose) & !(frequency_english %in% c("","na")) & insulin_type != "",.N]
```
When info on dose/frequency/type is recorded what are the various numbers?

Insulin type:

```{r}
insulin_orders[route == "sc",.N,by=insulin_acting][order(-N)]
```

Insulin dose (short):

```{r}
insulin_orders[insulin_acting == "short" & route == "sc",.N,by=dose][order(dose)]
```

Insulin frequency (short):

```{r}
insulin_orders[insulin_acting == "short" & route == "sc",.N,by=times_per_day][order(times_per_day)]
```

Insulin dose (long):

```{r}
insulin_orders[insulin_acting == "long"& route == "sc",.N,by=dose][order(dose)]
```

Insulin frequency (long):

```{r}
insulin_orders[insulin_acting == "long" & route == "sc",.N,by=times_per_day][order(times_per_day)]
```

## Any insulin

## Plots

```{r}
# example infusions
set.seed(12)
infusion_max <- insulin_infusions[source_infusion_drug == 1 & grepl("units/hr",drugname),
                                  .(icu_los_hours=max(icu_los_hours),insulin_rate=max(insulin_rate)),by=patientunitstayid]
samp10 <- sample(insulin_infusions[icu_los_hours > 24 & hospitalid == 420 & time < 24,patientunitstayid],10)
infusiondrug_samp <- insulin_infusions[source_infusion_drug == 1 & patientunitstayid %in% samp10]
infusiondrug_samp[,id := paste0("ICU stay - ",LETTERS[as.numeric(factor(patientunitstayid))])]

p1 <- ggplot(infusiondrug_samp,aes(x=time/60,y=insulin_rate)) +
  geom_segment(aes(xend=time_next/60,yend=insulin_rate)) +
  geom_step() +
  scale_x_continuous(breaks = seq(0,48,by=12)) +
  coord_cartesian(xlim=c(0,48)) +
  geom_rect(aes(xmin = time/60, xmax = time_next/60, 
                ymin = 0, ymax = insulin_rate), alpha = 0.7,
            fill="lightblue") +
  #scale_y_continuous(labels = label_comma(),n.breaks=6) +
  theme_bw(base_size=14) +
  facet_wrap(~id,ncol=2) +
  labs(x = "Time since ICU admission (hours)",y="Insulin infusion rate (units/hour)")+
  theme(panel.grid.minor = element_blank())
p1
# distribution of infusion rate
p2 <- ggplot(insulin_infusions[source_infusion_drug == 1 & insulin_rate > 0 & !is.na(diabetes)],
             aes(x=insulin_rate)) +
  geom_histogram(binwidth=1,col="white",fill="lightblue") +
  coord_cartesian(xlim=c(0,20)) +
  facet_wrap(~diabetes_f,ncol=1,scales="free_y") +
  scale_y_continuous(labels = label_comma(),n.breaks=6) +
  theme_bw(base_size=14) +
  labs(x = "Insulin infusion rate (units/hour)",y="Count")+
  theme(panel.grid.minor = element_blank())
p2
# distribution of infusion subcutaneous
dose_time_insulin <- insulin_orders[,.N,by=.(dose,times_per_day,insulin_acting)]
insulin_orders[,dose_g := cut(dose,c(0,5,10,20,Inf))]
tab <- insulin_orders[!is.na(dose_g) & !is.na(times_per_day),.N,
           by=.(dose_g,insulin_acting,times_per_day)][order(dose_g,insulin_acting,times_per_day)]
tab[,p := 100*round(N/sum(tab$N),3)]
tab
p3 <- ggplot(insulin_orders[route=="sc" & dose > 0],aes(x=dose_g,fill=insulin_acting)) +
  geom_bar(position="dodge",col="white") +
  scale_fill_manual(name="Insulin type",
                    labels=c("Short","Intermediate","Long"),
                    values = c("#ADD8E6","#E6ADD8","#D8E6AD")) +
  labs(x = "Units of insulin",y = "Count") +
  scale_x_discrete(labels=c("0-5","5-10","10-20","20+")) +
  scale_y_continuous(labels = label_comma(),n.breaks=6) +
  theme_bw(base_size=14)+
  theme(panel.grid.minor = element_blank())
p3
p4 <- ggplot(insulin_orders[route=="sc"],aes(x=times_per_day,fill=insulin_acting)) +
  geom_bar(position = position_dodge(preserve = "single"),col="white") +
  labs(x = "Injections per day",y = "Count") +
  coord_cartesian(xlim=c(0.5,6.5)) +
  scale_y_continuous(labels = label_comma(),n.breaks=6) +
  scale_x_continuous(breaks=1:7) +
  scale_fill_manual(name="Insulin type",
                    labels=c("Short","Intermediate","Long"),
                    values = c("#ADD8E6","#E6ADD8","#D8E6AD")) +
  theme_bw(base_size=14)+
  theme(panel.grid.minor = element_blank())
p4
dose_low_high <- melt(insulin_orders[route=="sc"][,.(dose_low,dose_high)])
p5 <- ggplot(dose_low_high,aes(x=variable,y=value)) +
  geom_boxplot(width=0.2,fill="gray90") +
  labs(x = "",y = "Insulin\ndose (units)") +
  scale_y_continuous(n.breaks=6) +
  scale_x_discrete(labels=c("Minimum","Maximum")) +
  theme_bw(base_size=14)+
  theme(panel.grid.minor = element_blank())
p5
# insulin by glucose
p6 <- ggplot(glucose[n==1 & !is.na(diabetes)],aes(x=glucose_bedside,y=insulin_dy1,col=diabetes_f)) +
  geom_smooth(method = 'gam',formula=y ~ s(x, bs = "cs"),method.args=list(family=binomial()),
              alpha=0.4) +
  scale_x_log10(limits=c(80,500),n.breaks=8) +
  scale_color_manual(values=c("#7CC0D7","#D7927C"),name="") +
  scale_y_continuous(limits=c(0,1),n.breaks=6) +
  labs(x = "Admission blood glucose (mg/dL)",y = "Probability of\nreceiving insulin") +
  theme_bw(base_size=14) +
  theme(panel.grid.minor = element_blank())
p6

layout <- matrix(c(1,1,1,1,5,
                   1,1,1,1,6,
                   2,2,3,4,6,
                   2,2,3,4,6),ncol=4)
gridExtra::grid.arrange(p1+labs(title="(A)"),
                        p2+labs(title="(B)"),
                        p3+labs(title="(C)"),
                        p4+labs(title="(D)"),
                        p5+labs(title="(E)"),
                        p6+labs(title="(F)"),
                        layout_matrix=layout)
p_out <- gridExtra::arrangeGrob(p1+labs(title="(A)"),
                       p2+labs(title="(B)"),
                       p3+labs(title="(C)"),
                       p4+labs(title="(D)"),
                       p5+labs(title="(E)"),
                       p6+labs(title="(F)"),
                       layout_matrix=layout)
ggsave(plot = p_out,filename = "../graphs/insulin.png",width = 10,height = 12)
```

## Descriptive stats

Number ICU stays receiving insulin

```{r}
insulin_pts <- unique(c(insulin_orders[,patientunitstayid],
                insulin_admin[,patientunitstayid]))
print("All")
length(insulin_pts)
print("Survived")
length(unique(c(insulin_orders[hosp_mort == 0,patientunitstayid],
                insulin_admin[hosp_mort == 0]$patientunitstayid)))
print("Died")
length(unique(c(insulin_orders[hosp_mort == 1,patientunitstayid],
                insulin_admin[hosp_mort == 1]$patientunitstayid)))
```


```{r}
# admission blood glucose
glucose[n_icu == 1,summary(glucose)]
glucose[n_icu == 1 & insulin == 1,summary(glucose)]
glucose[n_icu == 1 & insulin == 0,summary(glucose)]
```

```{r}
# diabetes
diabetes[,insulin := fifelse(patientunitstayid %in% insulin_pts,1,0)]
diabetes[diabetes == 1,.N,by=insulin]
diabetes[diabetes == 0,.N,by=insulin]
```
